Py: write 'Hamiltonian/Energies' as float for lattice models

+++ utils/afqmctools/utils/io.py
@@ -379,7 +379,7 @@ def write_model_hamiltonian(
         # write Energies!
         f.create_dataset(
             'Hamiltonian/Energies',
-            data = np.array([0, 0])
+            data = np.array([0., 0.], dtype=np.float64),
         )
 
         f.create_dataset(
